discord-bot: make very long replies work, usually programs
When asking Mistral to program something very long, e.g. "write a very
extensive Java program to implement a snake game", it will take a few
minutes (!) on my laptop then push the whole content in one shot. If the
content is more than maxMessage, discord would refuse it. Fix that.

There's also the problem when the backticks escaped content is more than
maxMessage. In this case it used to fail. Now it breaks the program in
the middle, which is bad, but still better than no content at all.

The next step is to inject 3 backticks so the content can be split
correctly.