Get rid of KnownLLMs
This greatly simplify the tests.

Get rid of Mistral specific encodings. Tool calling can now be done at
an higher abstraction level.

This causes /list_models to be broken.
Get rid of /usage to be replaced with a more generic implementation.
Update llama-server version.
Reduce the number of models tested.